+++ ContaBancaria.cpp
@@ -3,8 +3,8 @@
 
 
 ContaBancaria::ContaBancaria() {}
-ContaBancaria::ContaBancaria(const float saldo): saldo{saldo}{}
-ContaBancaria::ContaBancaria(const unsigned long numeroConta, const float saldo): numeroConta{numeroConta}, saldo{0}{
+ContaBancaria::ContaBancaria(const double saldo): saldo{saldo}{}
+ContaBancaria::ContaBancaria(const unsigned long numeroConta, const double saldo): numeroConta{numeroConta}, saldo{0}{
 }
 ContaBancaria::~ContaBancaria() {}
 
@@ -16,25 +16,18 @@ void ContaBancaria::setNumeroConta(const unsigned long novoNumero){
 	this->numeroConta = novoNumero;
 }
 
-float ContaBancaria::getSaldo() const{
+double ContaBancaria::getSaldo() const{
 	return this->saldo;
 }
 
-void ContaBancaria::setSaldo(const float valor){
+void ContaBancaria::setSaldo(const double valor){
 	if(valor < 0)
 		throw (int)1;
 	this->saldo = this->saldo + valor;
 }
 
-// std::string ContaBancaria::getNome() const{
-// 	return this->nome;
-// }
 
-// void ContaBancaria::setNome(const std::string nome){
-// 	this->nome = nome;
-// }
-
-void ContaBancaria::sacar(const float valor){
+void ContaBancaria::sacar(const double valor){
 	if(valor > this->saldo)
 		throw (int)1;
 	this->saldo = this->saldo - valor;
