+++ ContaBancaria.hpp
@@ -1,6 +1,10 @@
 #ifndef CONTA_BANCARIA_HPP
 #define CONTA_BANCARIA_HPP
 
+
+
+
+
 #include "Cliente.hpp"
 #include "Banco.hpp"
 
@@ -9,26 +13,23 @@ class ContaBancaria{
 
 	public:
 		ContaBancaria();
-		ContaBancaria(const unsigned long numeroConta, const float saldo);
-		ContaBancaria(const float saldo);
+		ContaBancaria(const unsigned long numeroConta, const double saldo);
+		ContaBancaria(const double saldo);
 		~ContaBancaria();
 
 		unsigned long getNumeroConta() const;
 		void  setNumeroConta(const unsigned long novoNumero);
 
-		float getSaldo() const;
-		void setSaldo(const float valor);	//atualiza o saldo!
+		double getSaldo() const;
+		void setSaldo(const double valor);	//atualiza o saldo!
 
-		//std::string getNome() const;
-		//void setNome(const std::string nome);
 
-		void sacar(const float valor); 
+		void sacar(const double valor); 
 		void gerarNumero();
 
 	private:
-		//std::string nome;
 		unsigned long numeroConta;	//numero realizado para transasções
-		float saldo;
+		double saldo;
 
 };
